Affirm that gitdb and smmap advisories can also be created
This expands `SECURITY.md` to affirm the claims in the new
`SECURITY.md` files in gitdb and smmap that vulnerabilities found
in them can be reported in the GitPython repository with the same
link as one would use to report a GitPython vulnerability, as well
as to note how the distinction between affected package can be
specified when it is known at the time a vulnerability is reported.

Along with https://github.com/gitpython-developers/smmap/pull/59
and https://github.com/gitpython-developers/gitdb/pull/117, this
fixes https://github.com/gitpython-developers/gitdb/issues/116.