Preserve quoted leading and trailing single-line whitespace
At least in a single line, whitespace in a double-quoted value in a
configuration file, like `name = " abc def "`, would presumably be
intended. This removes the `strip()` call that is applied to text
`ConfigParser` obtained by removing the double quotes around it.

This slightly refines the changes in #2035 by dropping the `strip()`
call while continuing to remove opening and closing double quotes.