Split Cygwin CI into two test jobs
One job is for all tests except the `performance` tests, while the
other job is for only the `performance` tests.

The idea is to decrease the total time it takes for all CI jobs to
complete in most cases, by splitting the long-running (currently
usually about 13 minute) Cygwin job into two less-long jobs.