__init__: don't run refresh on import
Consumers of gitpython may not need to use it in all use cases, and may
want to be able to run (without using gitpython) in environments where
git is not available on PATH. While this can be worked around by setting
the GIT_PYTHON_REFRESH environment variable, adding special handling for
gitpython means that it can't be imported just like everything else in
an import block at the top of the module, and environment variables have
potentially undesired propagation behaviour.

Previously, it was also nontrivial to distinguish gitpython failing to
import because of missing git or because e.g. gitpython isn't installed
at all, because the exception that's raised is an ImportError without
further qualification (except in the error message).

Thus, we now no longer perform `refresh` at the module top level,
instead performing it lazily when an invocation of git is attempted.
This also allows some functionality that doesn't rely on the git command
to work without, e.g. ref listing.